pyproject.toml
[tool.poetry]
name = "insta-sync"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
requests = "^2.32.5"
python-dotenv = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-django = "^4.12.0"
types-requests = "^2.32.4.20260107"
django-stubs = "^5.2.9"
djangorestframework-stubs = "^3.16.8"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.19.1"
black = "^26.1.0"
isort = "^8.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
